+++ lib/src/models/documents/document.dart
@@ -1,9 +1,7 @@
 import 'dart:async';
 
-import 'package:flutter_quill/src/widgets/sym_widgets/sym_title_widgets/sym_title.dart';
 import 'package:tuple/tuple.dart';
 
-import '../../../flutter_quill.dart';
 import '../quill_delta.dart';
 import '../rules/rule.dart';
 import 'attribute.dart';
@@ -198,9 +196,17 @@ class Document {
     return text;
   }
 
+  /// Only attributes applied to all characters within this range are
+  /// included in the result.
   Style collectStyle(int index, int len) {
     final res = queryChild(index);
     return (res.node as Line).collectStyle(res.offset, len);
+  }
+
+  /// Returns all styles for any character within the specified text range.
+  List<Style> collectAllStyles(int index, int len) {
+    final res = queryChild(index);
+    return (res.node as Line).collectAllStyles(res.offset, len);
   }
 
   ChildQuery queryChild(int offset) {
@@ -222,14 +228,16 @@ class Document {
   }
 
   void compose(Delta delta, ChangeSource changeSource,
-      {bool autoAppendNewlineAfterImage = true}) {
+      {bool autoAppendNewlineAfterImage = true,
+      bool autoAppendNewlineAfterVideo = true}) {
     assert(!_observer.isClosed);
     delta.trim();
     assert(delta.isNotEmpty);
 
     var offset = 0;
     delta = _transform(delta,
-        autoAppendNewlineAfterImage: autoAppendNewlineAfterImage);
+        autoAppendNewlineAfterImage: autoAppendNewlineAfterImage,
+        autoAppendNewlineAfterVideo: autoAppendNewlineAfterVideo);
     final originalDelta = toDelta();
     for (final op in delta.toList()) {
       final style =
@@ -274,37 +282,44 @@ class Document {
   bool get hasRedo => _history.hasRedo;
 
   static Delta _transform(Delta delta,
-      {bool autoAppendNewlineAfterImage = true}) {
+      {bool autoAppendNewlineAfterImage = true,
+      bool autoAppendNewlineAfterVideo = true}) {
     final res = Delta();
     final ops = delta.toList();
     for (var i = 0; i < ops.length; i++) {
       final op = ops[i];
       res.push(op);
       if (autoAppendNewlineAfterImage) {
-        _autoAppendNewlineAfterImage(i, ops, op, res);
+        _autoAppendNewlineAfterEmbeddable(i, ops, op, res, 'image');
+      }
+      if (autoAppendNewlineAfterVideo) {
+        _autoAppendNewlineAfterEmbeddable(i, ops, op, res, 'video');
       }
     }
     return res;
   }
 
-  static void _autoAppendNewlineAfterImage(
-      int i, List<Operation> ops, Operation op, Delta res) {
-    final nextOpIsImage =
-        i + 1 < ops.length && ops[i + 1].isInsert && ops[i + 1].data is! String;
+  static void _autoAppendNewlineAfterEmbeddable(
+      int i, List<Operation> ops, Operation op, Delta res, String type) {
+    final nextOpIsImage = i + 1 < ops.length &&
+        ops[i + 1].isInsert &&
+        ops[i + 1].data is Map &&
+        (ops[i + 1].data as Map).containsKey(type);
     if (nextOpIsImage &&
         op.data is String &&
         (op.data as String).isNotEmpty &&
         !(op.data as String).endsWith('\n')) {
       res.push(Operation.insert('\n'));
     }
-    // Currently embed is equivalent to image and hence `is! String`
-    final opInsertImage = op.isInsert && op.data is! String;
+    // embed could be image or video
+    final opInsertImage =
+        op.isInsert && op.data is Map && (op.data as Map).containsKey(type);
     final nextOpIsLineBreak = i + 1 < ops.length &&
         ops[i + 1].isInsert &&
         ops[i + 1].data is String &&
         (ops[i + 1].data as String).startsWith('\n');
     if (opInsertImage && (i + 1 == ops.length - 1 || !nextOpIsLineBreak)) {
-      // automatically append '\n' for image
+      // automatically append '\n' for embeddable
       res.push(Operation.insert('\n'));
     }
   }
@@ -335,8 +350,7 @@ class Document {
     assert((doc.last.data as String).endsWith('\n'));
 
     var offset = 0;
-    for (var i = 0; i < doc.toList().length; i++) {
-      final op = doc.toList().elementAt(i);
+    for (final op in doc.toList()) {
       if (!op.isInsert) {
         throw ArgumentError.value(doc,
             'Document can only contain insert operations but ${op.key} found.');