+++ lib/src/widgets/text_line.dart
@@ -27,6 +27,7 @@ class TextLine extends StatelessWidget {
     required this.line,
     required this.embedBuilder,
     required this.styles,
+    required this.readOnly,
     this.textDirection,
     Key? key,
   }) : super(key: key);
@@ -35,6 +36,7 @@ class TextLine extends StatelessWidget {
   final TextDirection? textDirection;
   final EmbedBuilder embedBuilder;
   final DefaultStyles styles;
+  final bool readOnly;
 
   @override
   Widget build(BuildContext context) {
@@ -46,7 +48,7 @@ class TextLine extends StatelessWidget {
     final childCount = line.childCount;
     if (line.hasEmbed || (childCount > 1 && line.children.first is Embed)) {
       final embed = line.children.first as Embed;
-      return EmbedProxy(embedBuilder(context, embed));
+      return EmbedProxy(embedBuilder(context, embed, readOnly));
     }
 
     final textSpan = _buildTextSpan(context);