+++ lib/src/widgets/toolbar.dart
@@ -2,11 +2,11 @@ import 'dart:io';
 
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:image_picker/image_picker.dart';
 
 import '../models/documents/attribute.dart';
 import 'controller.dart';
 import 'toolbar/arrow_indicated_button_list.dart';
+import 'toolbar/camera_button.dart';
 import 'toolbar/clear_format_button.dart';
 import 'toolbar/color_button.dart';
 import 'toolbar/history_button.dart';
@@ -17,6 +17,7 @@ import 'toolbar/link_style_button.dart';
 import 'toolbar/select_header_style_button.dart';
 import 'toolbar/toggle_check_list_button.dart';
 import 'toolbar/toggle_style_button.dart';
+import 'toolbar/video_button.dart';
 
 export 'toolbar/clear_format_button.dart';
 export 'toolbar/color_button.dart';
@@ -32,10 +33,12 @@ export 'toolbar/toggle_check_list_button.dart';
 export 'toolbar/toggle_style_button.dart';
 
 typedef OnImagePickCallback = Future<String?> Function(File file);
-typedef ImagePickImpl = Future<String?> Function(ImageSource source);
+typedef OnVideoPickCallback = Future<String?> Function(File file);
 typedef FilePickImpl = Future<String?> Function(BuildContext context);
 typedef WebImagePickImpl = Future<String?> Function(
     OnImagePickCallback onImagePickCallback);
+typedef WebVideoPickImpl = Future<String?> Function(
+    OnVideoPickCallback onImagePickCallback);
 
 // The default size of the icon of a button.
 const double kDefaultIconSize = 18;
@@ -76,8 +79,10 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
     bool multiRowsDisplay = true,
     bool showCamera = true,
     OnImagePickCallback? onImagePickCallback,
+    OnVideoPickCallback? onVideoPickCallback,
     FilePickImpl? filePickImpl,
     WebImagePickImpl? webImagePickImpl,
+    WebVideoPickImpl? webVideoPickImpl,
     Key? key,
   }) {
     final isButtonGroupShown = [
@@ -89,7 +94,8 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
           showColorButton ||
           showBackgroundColorButton ||
           showClearFormat ||
-          onImagePickCallback != null,
+          onImagePickCallback != null ||
+          onVideoPickCallback != null,
       showHeaderStyle,
       showListNumbers || showListBullets || showListCheck || showCodeBlock,
       showQuote || showIndent,
@@ -168,21 +174,30 @@ class QuillToolbar extends StatelessWidget implements PreferredSizeWidget {
             icon: Icons.image,
             iconSize: toolbarIconSize,
             controller: controller,
-            imageSource: ImageSource.gallery,
             onImagePickCallback: onImagePickCallback,
             filePickImpl: filePickImpl,
             webImagePickImpl: webImagePickImpl,
           ),
-        if (onImagePickCallback != null && showCamera)
-          ImageButton(
-            icon: Icons.photo_camera,
+        if (onVideoPickCallback != null)
+          VideoButton(
+            icon: Icons.movie_creation,
             iconSize: toolbarIconSize,
             controller: controller,
-            imageSource: ImageSource.camera,
-            onImagePickCallback: onImagePickCallback,
+            onVideoPickCallback: onVideoPickCallback,
             filePickImpl: filePickImpl,
-            webImagePickImpl: webImagePickImpl,
+            webVideoPickImpl: webImagePickImpl,
           ),
+        if ((onImagePickCallback != null || onVideoPickCallback != null) &&
+            showCamera)
+          CameraButton(
+              icon: Icons.photo_camera,
+              iconSize: toolbarIconSize,
+              controller: controller,
+              onImagePickCallback: onImagePickCallback,
+              onVideoPickCallback: onVideoPickCallback,
+              filePickImpl: filePickImpl,
+              webImagePickImpl: webImagePickImpl,
+              webVideoPickImpl: webVideoPickImpl),
         if (isButtonGroupShown[0] &&
             (isButtonGroupShown[1] ||
                 isButtonGroupShown[2] ||